Build: update config file for webpack
Added and configured the webpack plugins to generate a service worker
and manifest file.

Added CSS loader and babel modules to import css and transpile ES6 and
above to ES5 compliant code. n.b. even though I think I'm not going to
have to worry about this as internet explorer is now dead 🙏